requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --resolver=backtracking -o requirements.txt
annotated-types==0.7.0
    # via pydantic
arcgis==2.4.0
    # via hdx-scraper-eve (pyproject.toml)
attrs==25.3.0
    # via
    #   frictionless
    #   jsonlines
    #   jsonschema
    #   referencing
cachetools==5.5.2
    # via arcgis
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via frictionless
charset-normalizer==3.4.1
    # via requests
ckanapi==4.8
    # via hdx-python-api
click==8.1.8
    # via
    #   dask
    #   geomet
    #   typer
cloudpickle==3.1.1
    # via dask
contourpy==1.3.1
    # via matplotlib
cryptography==44.0.2
    # via
    #   arcgis
    #   pyspnego
    #   requests-kerberos
cycler==0.12.1
    # via matplotlib
dask==2025.2.0
    # via arcgis
decorator==5.2.1
    # via gssapi
defopt==6.4.0
    # via hdx-python-api
dnspython==2.7.0
    # via email-validator
docopt==0.6.2
    # via
    #   ckanapi
    #   num2words
docutils==0.21.2
    # via defopt
email-validator==2.2.0
    # via hdx-python-api
et-xmlfile==2.0.0
    # via openpyxl
fonttools==4.56.0
    # via matplotlib
frictionless==5.18.0
    # via hdx-python-utilities
fsspec==2025.3.0
    # via dask
geomet==1.1.0
    # via arcgis
gssapi==1.9.0
    # via
    #   pyspnego
    #   requests-gssapi
hdx-python-api==6.3.8
    # via hdx-scraper-eve (pyproject.toml)
hdx-python-country==3.9.0
    # via hdx-python-api
hdx-python-utilities==3.8.5
    # via
    #   hdx-scraper-eve (pyproject.toml)
    #   hdx-python-api
    #   hdx-python-country
humanize==4.12.1
    # via frictionless
idna==3.10
    # via
    #   email-validator
    #   requests
ijson==3.3.0
    # via hdx-python-utilities
inflect==7.5.0
    # via quantulum3
isodate==0.7.2
    # via frictionless
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jinja2==3.1.6
    # via frictionless
jsonlines==4.0.0
    # via hdx-python-utilities
jsonpath-ng==1.7.0
    # via libhxl
jsonschema==4.23.0
    # via
    #   frictionless
    #   tableschema-to-template
jsonschema-specifications==2024.10.1
    # via jsonschema
keyring==25.6.0
    # via arcgis
kiwisolver==1.4.8
    # via matplotlib
krb5==0.7.1
    # via pyspnego
libhxl==5.2.2
    # via
    #   hdx-python-api
    #   hdx-python-country
locket==1.0.0
    # via partd
loguru==0.7.3
    # via hdx-python-utilities
lxml==5.3.1
    # via arcgis
makefun==1.15.6
    # via hdx-python-api
markdown-it-py==3.0.0
    # via rich
marko==2.1.2
    # via frictionless
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.1
    # via arcgis
matplotlib-inline==0.1.7
    # via arcgis
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.6.0
    # via
    #   inflect
    #   jaraco-classes
    #   jaraco-functools
num2words==0.5.14
    # via quantulum3
numpy==1.26.4
    # via
    #   arcgis
    #   contourpy
    #   matplotlib
    #   pandas
    #   pylerc
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.5
    # via hdx-python-utilities
packaging==24.2
    # via
    #   dask
    #   matplotlib
pandas==2.1.4
    # via arcgis
partd==1.4.2
    # via dask
petl==1.7.15
    # via frictionless
pillow==11.1.0
    # via
    #   arcgis
    #   matplotlib
ply==3.11
    # via
    #   jsonpath-ng
    #   libhxl
pockets==0.9.1
    # via sphinxcontrib-napoleon
puremagic==1.28
    # via arcgis
pyarrow==19.0.1
    # via arcgis
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via frictionless
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pylerc==4.0
    # via arcgis
pyparsing==3.2.1
    # via matplotlib
pyphonetics==0.5.3
    # via hdx-python-utilities
pyshp==2.3.1
    # via arcgis
pyspnego==0.11.2
    # via
    #   arcgis
    #   requests-kerberos
python-dateutil==2.9.0.post0
    # via
    #   frictionless
    #   hdx-python-utilities
    #   libhxl
    #   matplotlib
    #   pandas
python-io-wrapper==0.3.1
    # via libhxl
python-slugify==8.0.4
    # via
    #   ckanapi
    #   frictionless
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via
    #   dask
    #   frictionless
    #   tableschema-to-template
quantulum3==0.9.2
    # via hdx-python-api
ratelimit==2.2.1
    # via hdx-python-utilities
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   arcgis
    #   ckanapi
    #   frictionless
    #   hdx-python-api
    #   libhxl
    #   requests-file
    #   requests-gssapi
    #   requests-kerberos
    #   requests-oauthlib
    #   requests-toolbelt
requests-file==2.1.0
    # via hdx-python-utilities
requests-gssapi==1.3.0
    # via arcgis
requests-kerberos==0.15.0
    # via arcgis
requests-oauthlib==2.0.0
    # via arcgis
requests-toolbelt==1.0.0
    # via arcgis
rfc3986==2.0.0
    # via frictionless
rich==13.9.4
    # via typer
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.10
    # via hdx-python-utilities
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
setuptools==77.0.3
    # via ckanapi
shellingham==1.5.4
    # via typer
simpleeval==1.0.3
    # via frictionless
simplejson==3.20.1
    # via ckanapi
six==1.17.0
    # via
    #   ckanapi
    #   pockets
    #   python-dateutil
    #   sphinxcontrib-napoleon
sphinxcontrib-napoleon==0.7
    # via defopt
stringcase==1.2.0
    # via frictionless
structlog==25.2.0
    # via libhxl
tableschema-to-template==0.0.13
    # via hdx-python-utilities
tabulate==0.9.0
    # via frictionless
tenacity==9.0.0
    # via hdx-python-country
text-unidecode==1.3
    # via python-slugify
toolz==1.0.0
    # via
    #   dask
    #   partd
traitlets==5.14.3
    # via matplotlib-inline
truststore==0.10.1
    # via arcgis
typeguard==4.4.2
    # via inflect
typer==0.15.2
    # via frictionless
typing-extensions==4.12.2
    # via
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typeguard
    #   typer
tzdata==2025.1
    # via pandas
ujson==5.10.0
    # via arcgis
unidecode==1.3.8
    # via
    #   libhxl
    #   pyphonetics
urllib3==2.3.0
    # via
    #   arcgis
    #   libhxl
    #   requests
validators==0.34.0
    # via frictionless
wheel==0.45.1
    # via libhxl
xlrd==2.0.1
    # via hdx-python-utilities
xlrd3==1.1.0
    # via libhxl
xlsx2csv==0.8.4
    # via hdx-python-utilities
xlsxwriter==3.2.2
    # via tableschema-to-template
xlwt==1.3.0
    # via hdx-python-utilities
